GitHub Action for publishing Swagger doc
